+++ community/packages/views/rsuite/src/viewWithCss.ts
@@ -1,14 +1,14 @@
 import {BiDi, createView} from '@react-form-builder/core'
-import {RsLocalizationWrapper} from './components/RsLocalizationWrapper'
+import {RsViewWrapper} from './components/RsViewWrapper'
 import {ltrCssLoader, rtlCssLoader} from './cssLoader'
 import {models} from './models'
 
 /**
  * An assembled set of rSuite components, ready to be passed as a property to the FormViewer.
- * This view contains CSS loaders.
+ * This view contains CSS loaders, localization, and light/dark theming.
  */
 export const viewWithCss = createView(models)
   // The following parameters are required for correct CSS loading in LTR and RTL modes
-  .withViewerWrapper(RsLocalizationWrapper)
+  .withViewerWrapper(RsViewWrapper)
   .withCssLoader(BiDi.LTR, ltrCssLoader)
   .withCssLoader(BiDi.RTL, rtlCssLoader)