+++ premium/examples/with-angular-forms/step4-designer/src/app/app.component.ts
@@ -5,6 +5,7 @@ import {createElement, ForwardedRef} from 'react';
 import {
   ltrCssLoader,
   RsLocalizationWrapper,
+  RsViewWrapper,
   rSuiteComponents,
   rtlCssLoader
 } from "@react-form-builder/components-rsuite";
@@ -20,7 +21,7 @@ const componentsMetadata = rSuiteComponents.map(definer => definer.build())
 const viewerComponents = componentsMetadata.map((componentMetadata) => componentMetadata.model)
 
 const builderView = new BuilderView(componentsMetadata)
-  .withViewerWrapper(RsLocalizationWrapper)
+  .withViewerWrapper(RsViewWrapper)
   .withCssLoader(BiDi.LTR, ltrCssLoader)
   .withCssLoader(BiDi.RTL, rtlCssLoader)