+++ premium/examples/with-formik/src/App.tsx
@@ -1,6 +1,6 @@
 import {
   ltrCssLoader,
-  RsLocalizationWrapper,
+  RsViewWrapper,
   rSuiteComponents,
   rtlCssLoader
 } from '@react-form-builder/components-rsuite'
@@ -28,7 +28,7 @@ const formStorage: IFormStorage = {
 const loadForm = () => formStorage.getForm('')
 
 const builderView = new BuilderView(componentsMetadata)
-  .withViewerWrapper(RsLocalizationWrapper)
+  .withViewerWrapper(RsViewWrapper)
   .withCssLoader(BiDi.LTR, ltrCssLoader)
   .withCssLoader(BiDi.RTL, rtlCssLoader)