+++ premium/examples/with-remix/with-remix-v2/app/components/builder.client.tsx
@@ -1,6 +1,6 @@
 import {
   ltrCssLoader,
-  RsLocalizationWrapper,
+  RsViewWrapper,
   rSuiteComponents,
   rtlCssLoader
 } from '@react-form-builder/components-rsuite'
@@ -28,7 +28,7 @@ const formStorage: IFormStorage = {
 const loadForm = () => formStorage.getForm('')
 
 const view = new BuilderView(components)
-  .withViewerWrapper(RsLocalizationWrapper)
+  .withViewerWrapper(RsViewWrapper)
   .withCssLoader(BiDi.LTR, ltrCssLoader)
   .withCssLoader(BiDi.RTL, rtlCssLoader)